Split client code and exports boilerplate
